Create a marker per vector
