+++ src/chart.ts
@@ -15,6 +15,7 @@ import helpers from './helpers/index.js'
 import datumDefaults from './datum-defaults.js'
 import datumValidation from './datum-validation.js'
 import globals from './globals.mjs'
+import { randomId } from './utils.mjs'
 
 export interface ChartMetaMargin {
   left?: number
@@ -82,8 +83,6 @@ export class Chart extends EventEmitter.EventEmitter {
 
   private readonly id: string
 
-  readonly markerId: string
-
   public meta: ChartMeta
   /**
    * options are the input options sent to function plot.
@@ -131,12 +130,9 @@ export class Chart extends EventEmitter.EventEmitter {
   constructor(options: FunctionPlotOptions) {
     super()
 
-    const n = Math.random()
-    const letter = String.fromCharCode(Math.floor(n * 26) + 97)
     this.options = options
-    this.id = letter + n.toString(16).substring(2)
+    this.id = randomId()
     this.options.id = this.id
-    this.markerId = this.id + '-marker'
     Chart.cache[this.id] = this
     this.linkedGraphs = [this]
     this.meta = {}
@@ -409,22 +405,6 @@ export class Chart extends EventEmitter.EventEmitter {
       .selectAll('.clip')
       .attr('width', this.meta.width)
       .attr('height', this.meta.height)
-
-    // marker clip (for vectors)
-    defs
-      .append('clipPath')
-      .append('marker')
-      .attr('id', this.markerId)
-      .attr('viewBox', '0 -5 10 10')
-      .attr('refX', 10)
-      .attr('markerWidth', 5)
-      .attr('markerHeight', 5)
-      .attr('orient', 'auto')
-      .append('svg:path')
-      .attr('d', 'M0,-5L10,0L0,5L0,0')
-      .attr('stroke-width', '0px')
-      .attr('fill-opacity', 1)
-      .attr('fill', '#777')
   }
 
   private buildAxis() {
